+++ makefile
@@ -1,5 +1,5 @@
-snake: dist/snake.o dist/position.o dist/game.o dist/pyClient.o
-	clang -lncurses dist/snake.o dist/position.o dist/game.o dist/pyClient.o $(shell python3-config --cflags --embed --ldflags) -o snake
+snake: dist/main.o dist/snake.o dist/position.o dist/game.o dist/pyClient.o
+	clang -lncurses dist/main.o dist/snake.o dist/position.o dist/game.o dist/pyClient.o $(shell python3-config --cflags --embed --ldflags) -o snake
 dist/snake.o: src/snake.c src/snake.h src/position.h
 	clang -c src/snake.c -o dist/snake.o
 dist/position.o: src/position.c src/position.h
@@ -8,6 +8,8 @@ dist/game.o: src/game.c src/game.h src/snake.h src/pyClient.h
 	clang $(shell python3-config --cflags --embed) -c src/game.c -o dist/game.o
 dist/pyClient.o: src/pyClient.c src/pyClient.h
 	clang $(shell python3-config --cflags --embed) -c src/pyClient.c -o dist/pyClient.o
+dist/main.o: src/main.c src/main.h src/game.h
+	clang $(shell python3-config --cflags --embed) -c src/main.c -o dist/main.o
 
 test: dist/test.o
 	clang dist/test.o -o test
