+++ src/main.h
@@ -1,0 +1,7 @@
+#ifndef main_h
+#define main_h
+
+#include "game.h"
+#include <curses.h>
+
+#endif
