+++ src/position.h
@@ -1,7 +1,6 @@
 #ifndef position_h
 #define position_h
 
-
 typedef struct {
     int x;
     int y;
@@ -14,7 +13,6 @@ typedef struct Corner{
     bool visited;
     Position* position;
 } Corner;
-
 
 Position* initPosition();
 Corner* initCorner(int x, int y, int dir);
